// nodes/HaloPSA/actions/interfaces/projects.ts
export interface HaloProjectsListResponse extends IDataObject {
	record_count: number;
	faults: HaloProjectSummary[];
	include_children: boolean;
}

export interface HaloProjectSummary extends IDataObject {
	id: number;
	dateoccurred?: string;
	summary?: string;
	details?: string;
	status_id?: number;
	status_name?: string;
	tickettype_id?: number;
	tickettype_name?: string;
	sla_id?: number;
	sla_name?: string;
	priority_id?: number;
	client_id?: number;
	client_name?: string;
	site_id?: number;
	site_name?: string;
	user_id?: number;
	user_name?: string;
	team_id?: number;
	team?: string;
	agent_id?: number;
	agent_name?: string;
	category_1?: string;
	category_2?: string;
	category_3?: string;
	category_4?: string;
	estimate?: number;
	estimatedays?: number;
	child_count?: number;
	attachment_count?: number;
	flagged?: boolean;
	read?: boolean;
	enduserstatus?: number;
	onhold?: boolean;
	respondbydate?: string;
	fixbydate?: string;
	excludefromsla?: boolean;
	slaholdtime?: number;
	site_timezone?: string;
	lastactiondate?: string;
	last_update?: string;
	organisation_id?: number;
	department_id?: number;
	reportedby?: string;
	user_email?: string;
	emailtolist?: string;
	emailcclist?: string;
	matched_kb_id?: number;
	product_id?: number;
	release_id?: number;
	release2_id?: number;
	release3_id?: number;
	lastincomingemail?: string;
	workflow_id?: number;
	workflow_step?: number;
	workflow_seq?: number;
	pipeline_stage_id?: number;
	is_vip?: boolean;
	isimportantcontact?: boolean;
	inactive?: boolean;
	impact?: number;
	urgency?: number;
	starttime?: string;
	starttimeslot?: number;
	targetdate?: string;
	targettime?: string;
	targettimeslot?: number;
	deadlinedate?: string;
	oppcompanyname?: string;
	oppvalueadjusted?: number;
	cost?: number;
	quantity?: number;
	userdef1?: string;
	userdef2?: string;
	userdef3?: string;
	userdef4?: string;
	userdef5?: string;
	source?: number;
	release_important?: boolean;
	releasenotegroup_id?: number;
	supplier_status?: number;
	appointment_type?: number;
	section_timezone?: string;
	projectinternaltask?: boolean;
	impactlevel?: number;
	guid?: string;
	reviewed?: boolean;
	merged_into_id?: number;
	table?: number;
	use?: string;
	maximumRestrictedPriority?: number;
	idsummary?: string;
	ticketage?: number;
	useful_count?: number;
	notuseful_count?: number;
	updateservicestatus?: boolean;
	servicestatusnote?: string;
	itil_requesttype_id?: number;
	ticket_tags?: string;
	invoiceseperatelyoverride?: boolean;
	purchaseordernumber?: string;
}